hip: fix error of ambiguous sqrt
It may reduce the precision 16 bytes to 8 bytes

+++ src/Healpix_2.15a/cxxutils.h
@@ -103,7 +103,7 @@ template<typename I> inline unsigned int isqrt (I arg)
 		return unsigned (sqrt(arg+0.5));
 	  else
 	  {
-		  long double arg2 = arg;
+		  double arg2 = arg;
 		  return unsigned (sqrt(arg2+0.5));
 	  }
   }
